V3 du composant PricingCard avec création des type et amélioration de la sémentique
